fix: change movie id type in movie translation; add movie year to search response

+++ src/main/java/com/retrolad/mediatron/model/movie/MovieTranslationId.java
@@ -15,7 +15,7 @@ import java.io.Serializable;
 public class MovieTranslationId implements Serializable {
 
     @Column(name = "movie_id")
-    private int movieId;
+    private Long movieId;
 
     @Column(name = "lang_code", length = 2)
     private String langCode;
